Add optional flags to clean distributions using remove command
