--- src/main/resources/cli-help/ballerina-dist-clean.help
@@ -1,9 +1,0 @@
-NAME
-       ballerina-dist-clean - Clean available distributions except active distribution.
-
-SYNOPSIS
-       ballerina dist clean
-
-DESCRIPTION
-       Clean removes the ballerina distributions from your local environment except the active distribution.
-
