+++ src/main/resources/cli-help/ballerina-dist-remove.help
@@ -2,17 +2,23 @@ NAME
        ballerina-remove - Remove distributions in your local environment
 
 SYNOPSIS
+       ballerina dist remove <-a | --all>
        ballerina dist remove <distribution>
 
 DESCRIPTION
-       Remove deletes the given Ballerina distribution from your local environment.
+       Remove deletes the given Ballerina distribution from your local environment if you specify the version.
 
        However, you are not allowed to remove the active distribution.
 
        Here, the "distribution" has two parts: name and version. Following are some sample
-       distributions: jballerina-1.1.0, jballerina-1.4.0, nballerina-0.5.0.
+              distributions: jballerina-1.1.0, jballerina-1.4.0, nballerina-0.5.0.
+
+       Remove deletes the all Ballerina distribution except currently active distribution from your local
+       environment if you issue <-a | --all> flag.
 
 
 EXAMPLES
        This command deletes jballerina-1.0.5 version in your local machine.
           $ ballerina dist remove jballerina-1.0.5
+       This command deletes all distributions except currently active distribution in your local machine.
+          $ ballerina dist remove -a
